Continue

+++ models/categories.model.js
@@ -1,22 +1,40 @@
 // Import Category Model Object
 
-const { categoryModel } = require("../models/all.models");
+const { categoryModel, adminModel, mongoose } = require("../models/all.models");
 
-async function addNewCategory(categoryName) {
+async function addNewCategory(authorizationId, storeIdAndCategoryName) {
     try{
-        const category = await categoryModel.findOne({ name: categoryName });
-        if (category) {
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const category = await categoryModel.findOne({ name: storeIdAndCategoryName.categoryName });
+                if (category) {
+                    return {
+                        msg: "Sorry, This Cateogry Is Already Exist !!",
+                        error: true,
+                        data: {},
+                    }
+                }
+                const newCategory = new categoryModel({
+                    name: storeIdAndCategoryName.categoryName,
+                    storeId: storeIdAndCategoryName.storeId,
+                });
+                await newCategory.save();
+                return {
+                    msg: "Adding New Category Process Has Been Successfuly ...",
+                    error: false,
+                    data: {},
+                }
+            }
             return {
-                msg: "Sorry, This Cateogry Is Already Exist !!",
+                msg: "Sorry, Permission Denied !!",
                 error: true,
                 data: {},
             }
         }
-        const newCategory = new categoryModel({ name: categoryName });
-        await newCategory.save();
         return {
-            msg: "Adding New Category Process Has Been Successfuly ...",
-            error: false,
+            msg: "Sorry, This Admin Is Not Exist !!",
+            error: true,
             data: {},
         }
     }
@@ -25,12 +43,12 @@ async function addNewCategory(categoryName) {
     }
 }
 
-async function getAllCategories() {
+async function getAllCategories(filters) {
     try {
         return {
             msg: "Get All Categories Process Has Been Successfully !!",
             error: false,
-            data: await categoryModel.find({}),
+            data: await categoryModel.find(filters),
         }
     }
     catch (err) {
@@ -64,23 +82,38 @@ async function getAllCategoriesInsideThePage(pageNumber, pageSize, filters) {
     }
 }
 
-async function deleteCategory(categoryId) {
+async function deleteCategory(authorizationId, categoryId) {
     try {
-        const category = await categoryModel.findOneAndDelete({
-            _id: categoryId,
-        });
-        if (category) {
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const category = await categoryModel.findOneAndDelete({
+                    _id: categoryId,
+                });
+                if (category) {
+                    return {
+                        msg: "Deleting Category Process Has Been Successfuly ...",
+                        error: false,
+                        data: {},
+                    };
+                }
+                return {
+                    msg: "Sorry, This Category Is Not Exist !!",
+                    error: true,
+                    data: {},
+                };
+            }
             return {
-                msg: "Deleting Category Process Has Been Successfuly ...",
-                error: false,
+                msg: "Sorry, Permission Denied !!",
+                error: true,
                 data: {},
-            };
+            }
         }
         return {
-            msg: "Sorry, This Category Is Not Exist !!",
+            msg: "Sorry, This Admin Is Not Exist !!",
             error: true,
             data: {},
-        };
+        }
     }
     catch (err) {
         throw Error(err);
@@ -89,19 +122,34 @@ async function deleteCategory(categoryId) {
 
 async function updateCategory(categoryId, newCategoryName) {
     try {
-        const category = await categoryModel.findOneAndUpdate( { _id: categoryId } , { name: newCategoryName });
-        if (category) {
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const category = await categoryModel.findOneAndUpdate( { _id: categoryId } , { name: newCategoryName });
+                if (category) {
+                    return {
+                        msg: "Updating Category Process Has Been Successfuly !!",
+                        error: false,
+                        data: {},
+                    };
+                }
+                return {
+                    msg: "Sorry, This Category Is Not Exist !!",
+                    error: true,
+                    data: {},
+                };
+            }
             return {
-                msg: "Updating Category Process Has Been Successfuly !!",
-                error: false,
+                msg: "Sorry, Permission Denied !!",
+                error: true,
                 data: {},
-            };
+            }
         }
         return {
-            msg: "Sorry, This Category Is Not Exist !!",
+            msg: "Sorry, This Admin Is Not Exist !!",
             error: true,
             data: {},
-        };
+        }
     }
     catch (err) {
         throw Error(err);
